+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.10
 - Implemented User registration (RestrrBuilder#register)
+- Added Restrr#createUser
 
 ## 0.9.1
 - Fixed use of `EntityId<E>`